# perf: make `ape --help` faster