… .pre-commit-config.yaml
… docs/methoddocs/ape.md
… docs/methoddocs/types.md
… src/ape/cli/options.py
… src/ape/managers/chain.py
… src/ape/managers/compilers.py
… src/ape/managers/config.py
… src/ape/managers/networks.py
… src/ape/managers/plugins.py
… src/ape/managers/query.py
… src/ape/plugins/compiler.py
… src/ape/plugins/config.py
… src/ape/plugins/converter.py
… src/ape/plugins/network.py
… src/ape/plugins/project.py
… src/ape/plugins/query.py
… src/ape/pytest/plugin.py
… src/ape/types/basic.py
… src/ape/utils/__init__.py
… src/ape/utils/basemodel.py
… src/ape/utils/misc.py
… src/ape/utils/trace.py
… src/ape_accounts/_cli.py
… src/ape_cache/__init__.py
… src/ape_cache/_cli.py
… src/ape_console/_cli.py
… src/ape_console/config.py
… src/ape_ethereum/__init__.py
… src/ape_ethereum/multicall/handlers.py
… src/ape_init/_cli.py