@fubuloubu 2024-10-24
I hate to say this, but with some of the "public" api moving around would it make sense to call this a breaking change?