@antazoey 2024-10-24
> I hate to say this, but with some of the "public" api moving around would it make sense to call this a v0.9 breaking change?

Like for example, if something was in `ape.types.__init__` and was moved to a submodule, the import from `ape.types` should still work for users. That was intent everywhere anyway.

Internally, we are importing from more direct sources but that is just for making Ape faster and not a clear indication of public API usage.

I am still testing all the plugins just to be sure, and can test ApePay as well.