@antazoey 2024-10-24
I did find something with ApePay but technically ApePay was already doing it wrong.
I can still re-expose the values from `ape.types` just in case but this never should have worked to begin with.
(albeit `ape.types` may be a better namespace for the base model stuff, maybe a 0.9-ism)

https://github.com/ApeWorX/ApePay/pull/115