@antazoey 2024-10-25
@fubuloubu OK, I have tested a bunch of stuff to ensure nothing is breaking anywhere...
I am still down for punting this to 0.9 if you want, but I feel like it is pretty safe. You can see everything I checked in this comment: https://github.com/ApeWorX/ape/pull/2333#issuecomment-2436050272

Also - down to check more repos and such too.

Thoughts?